refactor image source
